requirements.txt
Flask==1.0.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.3.1
flask-jwt-extended==3.13.1
